split users and authentications model test case from Accounts
